+++ test/src/conv.rs
@@ -141,3 +141,33 @@ pub fn array_conv(a: [u8; 5]) -> [u8; 7] {
     b[6] = 255;
     b
 }
+
+#[ocaml::func]
+#[ocaml::sig("'a -> ('a, 'b) Result.t")]
+pub fn result_ok(x: ocaml::Value) -> Result<ocaml::Value, ocaml::Value> {
+    Ok(x)
+}
+
+#[ocaml::func]
+#[ocaml::sig("'a -> ('b, 'a) Result.t")]
+pub fn result_error(x: ocaml::Value) -> Result<ocaml::Value, ocaml::Value> {
+    Err(x)
+}
+
+#[ocaml::func]
+#[ocaml::sig("('a, 'b) Result.t -> 'a option")]
+pub fn result_get_ok(x: Result<ocaml::Value, ocaml::Value>) -> Option<ocaml::Value> {
+    match x {
+        Ok(x) => Some(x),
+        Err(_) => None,
+    }
+}
+
+#[ocaml::func]
+#[ocaml::sig("('a, 'b) Result.t -> 'b option")]
+pub fn result_get_error(x: Result<ocaml::Value, ocaml::Value>) -> Option<ocaml::Value> {
+    match x {
+        Ok(_) => None,
+        Err(x) => Some(x),
+    }
+}